StockBot-Backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
